requirements.txt
pillow
pillow-heif
piexif
ffmpeg-python
geopy
requests